Move `CommentPresenter` to production
