+++ HackrNewsTests/Story Details Presentation/CommentPresentationTests.swift
@@ -5,89 +5,6 @@
 import HackrNews
 import XCTest
 
-struct CommentErrorViewModel {
-    let error: String?
-    static let none = CommentErrorViewModel(error: nil)
-}
-
-protocol CommentErrorView {
-    func display(_ viewModel: CommentErrorViewModel)
-}
-
-struct CommentLoadingViewModel {
-    let isLoading: Bool
-
-    static let loading = CommentLoadingViewModel(isLoading: true)
-    static let stoped = CommentLoadingViewModel(isLoading: false)
-}
-
-protocol CommentLoadingView {
-    func display(_ viewModel: CommentLoadingViewModel)
-}
-
-struct CommentViewModel {
-    let author: String
-    let text: String
-    let createdAt: String
-}
-
-protocol CommentView {
-    func display(_ viewModel: CommentViewModel)
-}
-
-class CommentPresenter {
-    private let calendar: Calendar
-    private let locale: Locale
-    private let view: CommentView
-    private let loadingView: CommentLoadingView
-    private let errorView: CommentErrorView
-
-    init(
-        view: CommentView,
-        loadingView: CommentLoadingView,
-        errorView: CommentErrorView,
-        calendar: Calendar = Calendar(identifier: .gregorian),
-        locale: Locale = .current
-    ) {
-        self.view = view
-        self.loadingView = loadingView
-        self.errorView = errorView
-        self.calendar = calendar
-        self.locale = locale
-    }
-
-    let errorMessage =
-        NSLocalizedString(
-            "loading_comment_error_message",
-            tableName: "StoryDetails",
-            bundle: Bundle(for: StoryDetailsPresenter.self),
-            value: "",
-            comment: "Comment loading error message"
-        )
-
-    func didStartLoadingComment() {
-        loadingView.display(.loading)
-        errorView.display(.none)
-    }
-
-    func didFinishLoadingComment(with comment: StoryComment) {
-        loadingView.display(.stoped)
-        view.display(CommentViewModel(author: comment.author, text: comment.text, createdAt: formatDate(from: comment.createdAt)))
-    }
-
-    func didFinishLoadingComment(with _: Error) {
-        loadingView.display(.stoped)
-        errorView.display(CommentErrorViewModel(error: errorMessage))
-    }
-
-    private func formatDate(from date: Date, against: Date = Date()) -> String {
-        let formatter = RelativeDateTimeFormatter()
-        formatter.calendar = calendar
-        formatter.locale = locale
-        return formatter.localizedString(for: date, relativeTo: against)
-    }
-}
-
 final class CommentPresentationTests: XCTestCase {
     func test_init_doesNotSendMessagesToView() {
         let (_, view) = makeSUT()
